images: Shrink windows-10 image and install QRes
Use sdelete to zero unused blocks, so that the image does not balloon as
much after installing updates.

Install QRes so that developers or our tests can switch to a higher
screen resolution programmatically. Set a default 1280x1024 resolution
which is a better start than the tiny 800x600 of the previous image.

Closes #32