Fix test/run
pyflakes or pycodestyle failures didn't cause the test to fail, as that
merely set `fail=1` which was then ignored.

We don't need TAP output here, so greatly simplify the structure and
rely on our `set -e`.

pycodestyle was called without any actual files to check (this was
already the case in the cockpit repo, bots/ was not covered by
pycodestyle). Fix the invocation, but disable the test for now as it
currently has hundreds of PEP-8 failures.