Rebase in make-checkout and not in tests-invoke
It makes much more sense to do this in make-checkout. The only problem
is that if rebasing fails, it won't be in logs. But since the goal is to
make from logging wrapper around all commands, then also make-checkout
would log what is happening.

Also it could be written as shell script - but that may be a follow-up.

Closes #121