images: Reduce libpwquality checks on ubuntu-stable
libpwquality 1.4.1 now also applies to `chpasswd`, breaking a lot of our
tests that want to set our standard "foobar" password. Reduce the
requirements, as that password is baked into our CI in so many places.

Same fix as in commit 2014019eebb086be for debian-stable; ubuntu-stable
is now 19.10, which has that libpwquality version.