images: Rename candlepin to services
Let's reduce the number of auxiliary images that we carry around and
have to download, and put the external services that tests talk to into
a single "services" image.

Start with candlepin, just rename the image scripts and adjust the
automatic rebuild. Also add an image refresh map, which was previously
missing for candlepin.

Do *not* remove images/candlepin just yet -- this needs to stay until
cockpit branches got moved to the new "services" image.