images: Drop fedora-29
It will be EOL in the next days, and we stopped releasing cockpit,
cockpit-podman, or any other project to Fedora 29.

Also stop testing cockpit on fedora-atomic, as it needs fedora-29 as a
build image, and Fedora Atomic becomes EOL at the same time. We soon
need to build and test a Fedora Core image instead.

Keep testing cockpit-ostree on fedora-atomic for now.