naughty: Consistent pattern for unexpected audit messages
The precise output format of unexpected messages has changed over time.
A long time ago with Pythoh 2 they were prefixed with `Error:`, then
with Python 3 it changed to `testlib.Error:`, and with [1] that prefix
moved to a separate line.

Change the naughties to accept any prefix, so that this works across all
branches.

[1] https://github.com/cockpit-project/cockpit/pull/13022

Closes #215