images: Grab EPEL 7 libssh on rhel-atomic image
Instead of downloading it through a container. That works fine on
continuous-atomic (which has its own setup script anyway), but not on
rhel-atomic any more, as that requires a valid subscription (which has
always been brittle). Use the EPEL 7 libssh instead, which is the same
version as in RHEL Atomic.

Do this with a direct download instead of pulling down an extra
container, and install it together with the cockpit packages.
This is much simpler, faster, and saves an extra 200 MB on the image
size.

This gets rid of the last bit in our bots that requires a subscription.