testmap: Drop cockpit rhel-8-appstream branch
This initially seemed like a good idea, but it's too much overhead to
maintain two branches per RHEL minor release, and does not actually buy
us anything. It's enough to maintain one branch, and cut releases from
it as we need them, as the resulting RPMs are not overlapping anyway.

Run rhel-8-1-distropkg on the rhel-8.1 branch instead.

Closes #414