machine: Drop VirtMachine boot diagnostics
This isn't really helpful. In CI it is responsible for a good deal of
long test run times when rebooting fails, and the diagnostics don't tell
us much about *why* a machine is completely frozen or *why* ssh to it is
broken even though it has a login prompt (the usual two cases that we
observe).

For a local developer, it's even less useful -- due to wrapping virsh
console into expect, the shown console does not react to keyboard input
at all, and furthermore blocks a manual `virsh console`, thus *breaking*
debuggability. It is much more helpful to run with --sit and attach to
the VM directly.

Fixes #507