images: Add cockpit/base podman container
On Fedora ≥ 31, use podman to build the cockpit/base container. This
will enable us to move the container/bastion test to Fedora 31.

Also make the "fedora:30" base image of cockpit/base dynamic, so that
building an image will automatically use a base image that corresponds
to the host Fedora version. With that we won't have to touch the
Dockerfile for future image upgrades any more, and the base container
will work on all our currently supported Fedora images. This requires
some sed hack for docker, but that will go away soon.

Drop the long-obsolete dnf upgrade workaround, that bug was fixed in
2017.

Closes #562