images: Move fedora-testing to Fedora 31
This was already intended in #501 and #567, but the symlinks were not
actually updated, so that the current image is still based on F30.

Also disable installing "atomic" on fedora-testing, for the same reason
as it's disabled on fedora-31. It's only a thing on fedora-30 now. We
still need to actually *install* docker so that we can install
cockpit-docker, we just can't use it.

There is no more "yum.conf" key in the mock config at last, so move to
"dnf.conf".

Fixes #570
Closes #569