images: Add ubuntu-2004
freeipa is not currently in the development series as it depends on a
broken dogtag-pki.

Don't copy ubuntu-stable naughties for now, tests seem fine. We can copy
them later on if still needed.

Add a naughty for Machines testCreate, as this keeps failing on our CI
(not locally). This will be investigated ASAP, but let's test all the
other bits until then.

Closes #612