images: Improve hack for pmlogger.service startup failure
Sometimes it takes pmlogger.service so long to start up that it misses
even the second attempt and times out again. Just disable the actual
ExecStart= command during the image build to avoid package installation
failure.

Note that masking does not work, as it will also confuse the postinst.

Fixes #679
Closes #727