machine: Fix Machine.start_cockpit() for cockpit/ws container
The container's `RUN` label intends to set the container name to "ws".
Current podman versions now actually respect that, but this caused
conflicts if tests call start_cockpit() multiple times without
stop_cockpit() in between. Since (1) this is all over the place, (2) for
rpm it does an implied stop first, and (3) it was always intended to
replace an existing container instead of starting a parallel one, just
call cockpit_stop() to clean up first.

Also fix stop_cockpit() to not fail if there is no running container.