naughty: Add pattern for podman coredumps
See known issue #805

In one case we get the actual traceback from `exe` where word
`rootlessport` seems to be the most important one.

In the other case there is not really much we can match to. The varlink
dies, thus disconnects us. Also we save coredump and it always fails
waiting for 'No Containers' string.