machine_core: Fix Machine.set_address()
Require the `nmcli con up` to succeed. It is absolutely not okay to
fail, as it will just cause weird test failures later on, due to
connection failures.

Robustify this by first stopping/deleting the interface's current
connection, so that the new one can succeed and is always picked.
Otherwise it happened in slow VMs that the "eth1" connection was still
active, and thus the provision's manual address never appeared.

Closes #877