image-create: Put VM image in /var/tmp/
At least in our CI, testvm.get_temp_dir() points to a tmpfs file, so
that tests can put VM overlays into RAM and are not affected by slow
I/O. But that doesn't work for image-create, as that's the entire image,
not just the overlay. This often caused the VM to crash due to ENOFS,
particularly during the "fill disk with zeros" phase.

Use /var/tmp/ instead, which is meant for large files.

Closes #912