images: Pre-install basic cockpit packages on Fedora images
This will make the images more efficient and robust for testing
starter-kit style cockpit extensions, as they don't need to invoke dnf
and package downloads any more.

It's also useful to have cockpit available for a quick interactive
"vm-run" exploration session, and for investigating regressions.

This uses the same approach as the rhel-* images have for a long time
already.