image-diff: new command for comparing images
This is a new command for displaying the differences in installed
packages and versions on two different images.

That's mostly useful when something broke because of an image refresh
and you'd like to find out what changed.

You'd usually want to use this with the full filename of a before- and
after-refresh version of the same image, for example:

  $ ./image-diff fedora-33-[oldhash].qcow2 fedora-33-[newhash].qcow2

It's also possible to use it on different images.  For example:

  $ ./image-diff rhel-8-3 rhel-8-4

Closes #1410