images: Move services image to Fedora CoreOS
Current FreeIPA containers are broken on CentOS/RHEL 7, and not really
supported by upstream any more:
https://github.com/freeipa/freeipa-container/issues/383

Unfortunately candlepin is *only* supported on RHEL/CentOS 7. To get out
of that deadlock, move candlepin into a CentOS 7 container, forward the
port, and provide a run-candlepin script that starts the container
(similar to run-{freeipa,samba}).  This is ultimately cleaner even, as
now *everything* on the services image is containerized.

With that, the host OS is much less important now, and we can use the
smallest one that we have -- so let's use Fedora CoreOS as container
host VM. This also significantly speeds up the image build. The main
drawback is that the /run-* scripts have to move to /root, as / is not
writable.

Adjust services.setup to get along with docker → podman, firewalld and
NetworkManager not being installed, a stricter SELinux profile (now
requiring `-v :z` option for container volumes), and the `admin` user
already existing (as it's created via Ignition already).