po-refresh: Don't fail when nothing to update
There are 3 cases we need to count with:
1. Some language was added/dropped
2. No language was added/dropped but some language was updated
3. Some language was added/dropped and some language was updated

All of these are handled correctly. However there is special case that
nothing was added/dropped nor there is nothing to update. In this case
it would fail, as `branch` would be `None` and when we would try to
create pull request we would get error message:
```
{'message': "Invalid request.\n\nFor 'properties/head', nil is not a string.", 'documentation_url': 'https://docs.github.com/rest/reference/pulls#create-a-pull-request'}
```
and we would try to get `number` property from this.

This was broken in 234053cfc4. (See the `if branch/elif local_branch)