task: Don't accept when PR fails to be created
This was introduced in https://github.com/cockpit-project/cockpit/commit/4c4160e87

The problem with the previous approach was that if the first `api.post`
would fail in any other way than with the `fork_collab` error we would end up
with unusable `pull` which would break something somewhere later on in
an unexpected way. (Like when we would try to read `number` or `url`).

Now we will see that PR failed to be created and also we will get some
human readable message:
```
po-refresh: Error accessing /repos/cockpit-project/cockpit/pulls
  Status: 422
  Reason: Unprocessable Entity
  Response: {"message":"Invalid request.\n\nFor 'properties/head', nil is not a string.","documentation_url":"https://docs.github.com/rest/reference/pulls#create-a-pull-request"}
```

before for the same cause we would get:
```
KeyError: 'number'
po-refresh: Task threw an exception
```
a few lines later on.

We might still have problem with the fork_collab but lets see how it
behaves and if we can address this in some other way (like editing
privileges...)