images: Build rhel-8-4 from official qcow image
Starting from the cloud image is faster, more robust, and more realistic
(using an official product) than our own virt-install bootstrap.

Teach cloudimage.bootstrap to get along with redirects, as
download.devel.redhat.com is a geolocation mirror.

The cloud images don't ship an empty /etc/crypttab, which uncovered a
bug in our storage page or in udisks:
https://github.com/cockpit-project/cockpit/issues/15100
Create an empy file in the install script to work around that for now.