npm-update: Restructure code for multiple updates
Pull the PR creation out of the update detection loop. Record updated
packages in an array, and prepare the PR creation for more than one
package. This isn't currently being used, but will soon.

Also drop the silent ignoring of task.branch() returning None; that only
happens if there was nothing to update, but with a non-empty
`updated_packages` this would be some bug which we rather want to know
through a crash than silently not creating PRs.

This also makes the code a bit easier to read.