machine: fix shell quoting on m.execute()
SSHConnection.execute() has existing support for passing the command as
a list instead of a string.  This seems like it might get a
subprocess-like interface for protecting you from the shell, but in
practice the list items are just passed as separate commandline
arguments to ssh, which will concatenate them and run them as a shell
command anyway.

Fix that by using shlex.join() to properly escape each string, and pass
the result as a single argument to ssh.

This behaviour was previously undocumented, so add it to the docstring.