task: Drop user from branch()/pull()
Commit 1b5a07419eb23e7 dropped using forks for our bots. After that, we
can simplify task.branch() to not parse and return the user name any
more, as it will always be the origin.

Most bots just directly pass the return value to task.pull(), adjust
that to not expect an "username:" prefix any more. po-refresh does some
more processing of the result, adjust that accordingly.