tests-scan: Use direct test trigger without --amqp
The primary way of calling tests-scan is through run-queue: This will
supply an --amqp server and queue the tests by itself. This uses status
`github.NOT_TESTED`, which run-queue will ignore (to avoid
triggering/queueing them again).

However, we also occasionally call tests-scan by hand, for examaple when
removing the no-test label from a PR. That does not really work, as the
tests will then never get into the queue. So far the workaround was to
run an additional `tests-trigger --force` on the PR. But we don't really
need this extra roundtrip -- when calling tests-scan without `--amqp`,
we *always* need something that flips the NO_TESTED to
NOT_TESTED_DIRECT, so tests-scan may just as well do that on its own.

This will also allow us to move test triggering from a webhook to a
workflow, as we plan to do for cockpit soon.