testmap: Add testmap injection
This is useful for cockpituous' self-tests so that they can use the same
tests-scan workflow as production [1], without having to play extra tricks
with tests-scan. That approach would not work at all for writing similar
integration self-tests for bots, as doing tests-trigger races with the
*real* bots (the bots repo has an issues webhook trigger, unlike
cockpituous).

This replaces the `_manual` trigger for cockpituous.

[1] https://github.com/cockpit-project/cockpituous/pull/376