weather: Parse number of days as integer
JS is amazing with dealing with this kind of situations:
```
> "5" +1
"51"

> "5" * 10
50
```

This manifested itself on the page - I was using `days` with just
multiplication and ti worked fine, but there is now also addition and it
does not behave as expected. Reproducible by changing number of days and
then limit lines in graphs disappear.