weather: Fix collection of statuses
We need to parse context as it may contain @bots or @project parts.
But we cannot do this straight away, as that is our only real unique
identifier.

If (like in https://github.com/cockpit-project/bots/pull/1637) we have
the same context name in two different projects, then the behavior is
hardly defined. As the logic would create item for example for d-testing
on cockpit, but then status for d-testing on c-podman would write there
its own data.