weather: Span missing graph points
If we don't merge anything on some day, graphs are becoming rather ugly.
With this option lines are drawn between points. No point is visually
drawn in the missing day and also we indicate in label that there was no
run.

It would be nice if https://github.com/chartjs/Chart.js/issues/5956
would be implemented, but in the meantime I guess we can just use this.
(Other option is to create for each graph two data sets, but that is rather busy work)