testmap: remove rhel-8-4-distropkg from master
We're about to introduce a change to cockpit which adds a variable
expansion feature to manifest parsing, and we want to use this feature
right away from the manifests in the same version.  This means that
the -distropkg variant is going to be broken until that new version gets
packaged.

Meanwhile, we're about to move over to RHEL 8.5 anyway.

It seems the easiest thing for now is to just drop rhel-8-4-distropkg
from the testmap.