task: move github-token to ~/.config/cockpit-dev
Similar to the last commit, ~/.config/github-token isn't sufficiently
namespaced.  Move it into cockpit-dev as well.

Unlike the last commit, we provide some backwards compatibility here.
This feature is very widely used and we don't want to suddenly break it
for everyone.

The excessive indentation here is pretty unfortunate, but python doesn't
really give a better way to do this pattern without checking for the
existance of the files first (which, itself, would be unpythonic).  This
extra code will evetually be removed.