lib: new file 'stores.py'
This file contains the list of the image stores, the path to the CA
certificate used to sign the certificate of the image stores, and a
utility function which checks if the redhat-internal image stores are
available.

All of those things have been removed from task/__init__.py.

Nobody is using these outside of the bots, so no compatibility shim for
this one.