lib: split stores into a text file
Break the list of the stores out of lib/stores.py and into a normal text
file at lib/stores.  Each line contains two fields, according to python
str.split(): the first is the "scope", which should either be "public"
or "redhat", and the second is the url.

This change will benefit cimg[1] which will no longer need to maintain
its own list.

[1] https://github.com/allisonkarlitskaya/cimg