image-download: lift "OpenShift" case out of loop
The "OpenShift" case where we just directly contact the server (albiet
with a custom CA certificate) doesn't use the IP address of the server,
so it makes no sense to have it in the loop that iterates over the
different IP addresses: the exact same command would be invoked each
time, irrespective.

Also: move it to be the first case we try, since it's the one that's
closer to "normal", and if it works, we can avoid the lookup mechanics
entirely.