image-{up,down}load: add support for S3 stores
Introduce a new "hybrid" type of store which can host either private or
public images.  This is intended to be used with an S3 bucket which is
configured to be restrictive by default.

When the image is uploaded we set the ACL to 'public_read' if the image
is a public image.  That means that downloading the image from the store
can be done by anyone.

Otherwise, we have a new lib/s3.py file which contains a small helper
for creating a signed version of the URL.  That is used both for
uploading (since writing must always be authenticated) and also for
downloading private images.  We also include an attempt to upload and
download unsigned versions of URLs without using our CA_PEM.

The code looks in ~/.config/cockpit-dev/s3-keys/[endpoint] to find the
user's key for a given endpoint.

So far, no actual S3 servers are configured.