test: Check naughty pattern matching with zero or more optional lines
Our naughties commonly use this pattern:

    some expected log message
    *
    Traceback..

But this does not match if there are *no* lines between the log message
and the Traceback. This is because the fnmatch still considers the (not
quite obvious) \n line break, and thus a log like

    blabla
    some expected log message
    Traceback foo..
       File ...

does *not* match the above pattern. The solution is to put the asterisk
at the end of the previous line:

    some expected log message*
    Traceback..

which will then correctly match zero or more lines in between.

Add a unit test to illustrate and validate this.