naughty: Fix patterns with unsafe "*" lines
As explained by the previous commmit, these patterns contained "*" lines
which could potentially happen with nothing in between the preceeding
and succeeding line. Rewrite these to put the asterisk at the previous
end of line, so that the pattern will accept "zero or arbitrarily many
lines in between".

This fixes the current fedora-34 TestAD.testQualifiedUsers failure.