stores: drop the internal-images mirror
Drop internal-images.cockpit-project.org.

This mirror only existed as a backup for rhel images incase the e2e
cluster went offline, and for that we now have the S3 mirrors.  It also
suffered from the name being resolvable to an internal IP from outside
the VPN leading to having to wait for the timeout (...as 10.x got
blackholed) before image-download could proceed, when not connected to
VPN.