image-download: add "preferred image store" support
image-download now looks in ~/.config/cockpit-dev/image-store-preference
on each download.  If this file exists, it should contain a substring
which matches the name of a single mirror (for example "eu-central-1").
In that case, this mirror will be tried before the other mirrors, and if
the image is present on it, the other mirrors will be completely
ignored.