image-upload: only try curl once
Instead of our mix-and-match approach of S3 or token authentication plus
https with and without a custom CA, simplify the logic: we know that S3
always works without a custom CA, and github token auth only works with.

Add an explicit error case for when we have no authentication, instead
of just trying to upload unauthenticated.

We now only ever call curl once, so stop using a nested function for
that.