s3: rework main()
Port the "get" command in main() over to using the new v4
request-signing scheme.  Add an "rm" command which invokes DELETE (which
is unsupported under the v2 URL-signing scheme).

Clean up and consolidate a bit of logic, and do proper shell quoting on
the output.

Add a comment as a pointer for how to actually use the cli.

The "put" and "put-public" commands are still using the url-signing
scheme because it lets us avoid thinking about checksums (which are
mandatory under v4 request-signing).