image-prune: improve detection of 'original' links
image-prune tries to make a distinction between symlinks with names like
'fedora-33' and 'fedora-33-[checksum].qcow2'.  It currenty does that by
checking for a '.qcow2' or a '.partial' extension, and treating those
files as non-"original".

I'm not sure why '.partial' is considered here, since as far as I know
we never create such links, but a much easier way would just be to check
if the filename has any dot ('.') in it at all, since no original image
name does.