image-prune: major object-oriented refactor
Cleanly separate the code for building the list of images to keep
("keepers") from the code for scanning, deciding, and removing files
from the image cache (ie: "the images data dir").

Port the later to an object-oriented approach with the generic decision
functionality in a parent class and the listing and removal
functionality split into a subclass.

This will allow the future addition of different kinds of image caches.

Also add a --debug commandline option for providing information about
why images are not pruned.