image-prune: add support for pruning S3 stores
Add a --s3 argument to allow specifying the URL of an S3 store to prune.
By default, S3 stores are considered "low on space" if they have more
than 80GB of images.

Add a bonus --directory argument for the sake of symmetry to allow
specifying a non-default local cache directory for pruning.