prometheus-stats: Export queue time histogram of last 12 hours
Looking back just one hour would give us zeros most of the time, because
we usually don't land several PRs within one hour, and there is
apparently some time zone shift. So extend the lookback time to 12
hours.

Also, we are not really interested in the average queue time, but how
many tests (absolute and percentage) have to wait longer than 5 mins
(our SLO) or 1 hour (large congestion), or longer (infrastructure is
down). So export the times in three buckets as a histogram.