prometheus-stats: Export unstable tests
Add a new top_failures_pct{project,test} metric for the topmost tests
that fail more than 10% of the time. These represent SLO violations, and
we want to expose them on Grafana.

This is still a bit experimental -- in particular, this does not yet
expose example URLs for further investigation.