task: Allow external link patterns for log.html
The idea is that testlib.py attaches an appropriate link-patterns.json
file for all the kinds of files that it attaches.  This will be used
by the pixel tests to make links to the pixels comparison pages.

Also, the default pattern for a screenshot is more specific so that we
don't bogusly pick up all the PNGs from the "initialisation" segment.