prometheus-stats: Move from 24h windows to indefinite counters
Only exporting data from the last 24 hours was a wrong design choice:
PromQL Data queries are a lot more flexible when exporting the data
since the beginning of time as counters instead of momentary gauges, and
selecting the desired time window on the PromQL/Grafana side.

That way, we can have different views, like the current 24 hour "short
time development" for seeing new or fixed errors; but also an 28 day
long-term view for seeing how much we are eating into our error budget,
or any other window.

The only remaining gauges are the set of >10% test failures from the last
7 days and the S3 usage, as these are both genuine gauges.

Fixes #1910