fedora.setup: force mock selinux-policy version == host
We're occasionally seeing situations where our mock chroot ends up with
a newer selinux-policy version than that which is installed on the host.

That's a problem, because cockpit now depends on (at least) the
selinux-policy version that was available at build time.  That means,
that if mock has a newer one, we won't be able to install the package on
the system.

We can work around this issue by making use of the
fedora-updates-archive repo (introduced for rpm-ostree to solve similar
versioning conflict problems) to ensure that we get our exact version of
selinux-policy into mock.