task: Fix is_user_allowed() for low-privileged tokens
If the token does not have `public_repo` permissions, reading the
project collaborators will fail. This caused a `GitHubError` crash
and prevented the intended fallback to self.allowlist().

Remove the collaborator check and depend entirely on the allowlist,
which should be readable with a relatively innocent `read:org` token.

Add some extra output to the exception printout in the case of 401
errors, directing the user to a freshly-updated section in README.md
about how to create a proper token.