github-upload-action-secrets: Upload to environment
Add a new `--env` option to upload the given secrets to a project
environment, instead of the organization or global project secrets. This
is useful for partitioning secrets and thus limiting their potential
exposure/leakage.

There is no such thing as organization environments, they are always
per-project (as of today). Thus change the default value of `--receiver`
to depend on whether or not `--env` is given. Also fix the case where
there is no repo organization (as in personal forks), just use the repo
itself as a fallback.

Also add a `--dry-run` option to see exactly what happens without
actually writing to the API.