machine: Don't send invalid DHCP hostname with colons
`control` is "127.0.0.1:<port>". Using that as DHCP-announced host name
is invalid, as `:` must not appear  in a hostname, they caused
NetworkManager/hostnamed to groan:

    policy: set-hostname: current hostname was changed outside NetworkManager: 'localhost'
    policy: set-hostname: set hostname to 'rhel-8-4-127.0.0.2\0582201' (from DHCPv4)
    hostname: couldn't set the system hostname to 'rhel-8-4-127.0.0.2\0582201' using hostnamed:
      GDBus.Error:org.freedesktop.DBus.Error.InvalidArgs: Invalid hostname 'rhel-8-4-127.0.0.2\0582201'

Replace the colon with a dash. Also rename the "name" key to "hostname"
to make it more obvious what this is for.

This has only a visible effect on the RHEL 8 image, as that prefers the
DHCP provided name over /etc/hostname's "localhost". This uncovered a
bug on that image: Some tests rely on being able to resolve its own
hostname, but `myhostname` appears after `dns` in nsswitch.conf, which
causes resolving the hostname to hang. Flip the resolution order to
avoid that.

Fixes #1908