machine: Disable DNS search domain
This avoids bogus options like `search 0.0.2\0582201` in
/etc/resolv.conf. Moreover, it also (somewhat mysteriously) fixes QEMU's
DHCP to send DNS servers with `restrict=on`. This caused long delays
with trying to resolve one's own hostname, as nsswitch.conf usually puts
`myhostname` after `dns`. This is a debatable design decision, but not
outright wrong. But it is wrong to advertise a DNS server which you
can't actually talk to, so let's fix that.