publish-wrapper: Stop sending API tokens through the sink protocol
The sink already has ~/.config/github-token, and falls back to that if
the request does not include it [1]. As in our practical purposes that
is the same token, drop it.

This avoids sending the token through the wire protocol all the time,
where it is very prone to getting leaked through error logging.

[1] https://github.com/cockpit-project/cockpituous/blob/5221fac7e/sink/sink#L149