tests-invoke: Avoid test failure due to GitHub API timeout during collision detection
During times when the GitHub API is a bit flaky (like right now), a lot
of tests fail with "TimeoutError: [Errno 110] Connection timed out" in
detect_collisions(). This is awkward -- it's not that bad if two tests
run in parallel for a little longer when skipping a collision check or
two. So ignore connection timeouts for that particular check.

Note that when the GitHub API is permanently down, test invocation and
result reporting will fail anyway, so it's not completely hidden.