machines: add back a graphics adaptor
We ripped this code out when removing our windows image, back in
b4ac348460bc63088a9588b7b87fae24de0b6e3a.  It turns out that this is
needed for `vm-run -G` to work (which still exists).  It's also useful
for debugging in some cases.

I considered enabling this conditionally, but it comes with very little
cost, so let's just turn it on by default.

It should now be possible, on any running VM to find the vm via `virsh
list` and connect to it, by ID number, with `virt-viewer`.  That already
came in useful for debugging issues while adding the new Arch Linux
image.