fedora.setup: stop relying on updates-archive
We can ensure we get the same versions of selinux-policy and friends
inside and outside of the container by simply downloading them with dnf
and installing them from the local .rpm files.

That lets us skip using the updates-archive repository which isn't yet
available for Fedora 35.

This is what we're already doing on rhel.