testmap: Cleanup unneeded contexts for weldr/lorax
Answer that I got from lorax maintainers:
```
I think we should leave only 'rhel-8-5/osbuild-composer' however I'm not
sure if that is ready to be executed automatically or no.

This component is scheduled to be replaced and we aren't paying much
attention to the old code anymore. CI will probably fail but we'll look
into lorax-composer only when there are bugs/PRs against this component.
```

If `rhel-8-5/osbuild-composer` works for `rhel8-branch` or `master` lets
test that, otherwise let's not bother if the team cannot be bothered
either :)