fedora.setup: use old selinux-policy in mock
Disable the fedora-updates repository while downloading the
selinux-policy packages to install into the mock.  The host gets the
normal version.

This ensures that the version of selinux-policy in the mock is sure to
be much older than this host or any other host (ie: fedora-coreos) that
we may try to install the cockpit-ws package on, meaning that our >=
depend on the selinux-policy version that we built against is sure to be
satisfied.