task/github: add "remote" property
In case the GitHub repository has been specified via GITHUB_BASE or the
git-config variable, this will return a https URL according to the
specified value.  Otherwise it will return "origin".

This will help us decide where to push things to in order to open a PR:
it may be the case that we have `origin` via SSH rather than via HTTPS,
and in that case, we may have no GitHub token (or an insufficiently
powerful token) configured for HTTPS.