npm-update: use `npm outdated` instead of `upgrade`
There's a much easier way to figure out which versions are available:
`npm` now features an `outdated` command which will tell us which
versions we'd end up with for a given version string.  Now we prefix all
the versions, run the tool to get the upgrades, and then apply our logic
for grouping and removing already-pending updates.