images: Upgrade debian-testing image to actual testing
There are no cloud images for bookworm yet (the future release which is
currently testing), so the debian-testing image is still being
bootstrapped from bullseye (current stable). Update the image to actual
testing again.

Drop some naughties which are fixed. But add pattern for issue #1739
which affects Debian testing as well now.