tests-scan: More precise filtering of existing contexts
When invoking tests-scan with -c to only operate on an existing status
(aka. context), we usually intend a precise matching of the entire
string, instead of just the OS/scenario part. This is true both for a
"statuses" webhook event (in run-queue) as well as manually calling
tests-scan for debugging.

Simplify the "existing statuses" filtering to do an exact match.
This was the only consumer of the optional "contexts" argument of
is_valid_context(), so drop it.