Eliminate testmap.known_context()
This is almost useless, as it entirely disregards a repo/branch
specification in a context, and only checked if that OS/scenario existed
in *any* project/branch.

Use the much more precise is_valid_context() in tests-trigger instead.
This will prevent us from manually triggering bogus/undeclared contexts
on a PR.

Also port GitHub.statuses() to is_valid_context(). The validation is not
that important there, it just needs to tell whether a status is for
Cockpit's CI or a third-party one. But being precise there cannot hurt
either.

Drop the now unused known_context() function.