tests-trigger: support specifying '-' as PR number
In case '-' is given as the PR number to trigger, look for the upstream
branch of the current branch, and trigger the tests on the corresponding
HEAD commit.

This will do the "right thing" most of the time for when you want to
trigger tests for the feature branch you're hacking on and just pushed
changes to.

This skips the "is the PR author trusted?" checks because we don't look
at the PR.  That's fine: this is the code you're working on locally.